feat(inventory): support decimal quantities for unit-of-measure billing
Migrate quantity/stock columns from Int to Decimal(10,3) across products,
warehouse stock, transfers, purchase/supplier/customer orders and all
billing item lines (sale, invoice, credit/debit note) so weighable and
measured products can be sold and tracked in fractional units. Adapt
backend services/routes/validation and frontend inventory, fiscal and
commercial views to the Decimal type.

+++ backend/src/routes/validities.ts
@@ -14,7 +14,7 @@ const router = Router();
 
 const createSchema = z.object({
     lote: z.string().max(100).optional().nullable(),
-    quantity: z.number().int().min(0),
+    quantity: z.number().min(0),
     expiryDate: z.string().min(1, 'Data de validade obrigatória'),
     costPrice: z.number().nonnegative().optional(),
     notes: z.string().max(500).optional().nullable(),
